Implement jitter for chaos monkey functions.

+++ pkg/chaos/crash.go
@@ -33,7 +33,11 @@ type CrashMonkey struct {
 	config  *v1alpha1.CrashMonkey
 }
 
-func (m *CrashMonkey) run(stop <-chan struct{}) {
+type CrashContainerMonkey struct {
+	*CrashMonkey
+}
+
+func (m *CrashContainerMonkey) run(stop <-chan struct{}) {
 	selector := labels.SelectorFromSet(util.NewClusterLabels(m.cluster))
 
 	listOptions := client.ListOptions{
@@ -58,12 +62,49 @@ func (m *CrashMonkey) run(stop <-chan struct{}) {
 	pod := pods.Items[rand.Intn(len(pods.Items))]
 	container := pod.Spec.Containers[0]
 
-	log.Info("Killing pod", "pod", pod.Name, "namespace", pod.Namespace)
+	log.Info("Killing container", "pod", pod.Name, "namespace", pod.Namespace, "container", container.Name)
 
 	// Kill the node by killing the Java process running inside the container. This forces health checks to be used
 	// to recover the node.
 	_, err = m.context.exec(pod, &container, "bash", "-c", "kill -9 $(ps -ef | grep AtomixAgent | grep -v grep | cut -c10-15 | tr -d ' ')")
 
+	if err != nil {
+		m.context.log.Error(err, "Failed to kill container", "pod", pod.Name, "namespace", pod.Namespace)
+	}
+
+	<-stop
+}
+
+type CrashPodMonkey struct {
+	*CrashMonkey
+}
+
+func (m *CrashPodMonkey) run(stop <-chan struct{}) {
+	selector := labels.SelectorFromSet(util.NewClusterLabels(m.cluster))
+
+	listOptions := client.ListOptions{
+		Namespace:     m.cluster.Namespace,
+		LabelSelector: selector,
+	}
+
+	// Get a list of pods in the current cluster.
+	pods := &v1.PodList{}
+	err := m.context.client.List(context.TODO(), &listOptions, pods)
+	if err != nil {
+		m.context.log.Error(err, "Failed to list pods")
+	}
+
+	// If there are no pods listed, exit the monkey.
+	if len(pods.Items) == 0 {
+		m.context.log.Info("No pods to kill")
+		return
+	}
+
+	// Choose a random node to kill.
+	pod := pods.Items[rand.Intn(len(pods.Items))]
+
+	log.Info("Killing pod", "pod", pod.Name, "namespace", pod.Namespace)
+	err = m.context.client.Delete(context.TODO(), &pod)
 	if err != nil {
 		m.context.log.Error(err, "Failed to kill pod", "pod", pod.Name, "namespace", pod.Namespace)
 	}
